pyproject.toml
[tool.poetry]
name = "iiwa-batter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "iiwa_batter"}]
package-mode = true

[tool.poetry.dependencies]
python = "3.11.10"
drake = "^1.32.0"
manipulation = "^2024.6.29"
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
nbconvert = "^7.16.4"
pydot = "^3.0.1"
ruff = "^0.1.5"
pre-commit = "^3.5.0"
jax = "^0.4.35"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true
select = [
  "A", # avoid shadowing
  "B", # flake8-bugbear
  "C4", # comprehensions 
  "E", # pycodestyle Errors
  "ERA", # no commented out code
  "F", # pyflakes
  "FLY", # flynt
  "I001", # isort
  "ISC", # implicit string concatenation
  "PERF", # Perflint
  "PIE", # flake8-pie
  "PGH", # pygrep-hooks
  "PL", # pylint
  "Q", # flake8-quotes
  "RUF", # ruff builtins e.g. noqa checking
  "T10", # flake8-debugger (no breakpoint etc)
  "TCH",# type-checking imports
  "UP", # pyupgrade
  "W", # pycodestyle warnings
  ]

ignore = [
  "E501", # Never enforce line length violations, we have black for that.
  "PLR0913", #ignore limit on number of args
  "PLR2004", #ignore magic values warning, at least for now
  "C408", # use {} instead of dict(), but we use dict heavily, for now leave it
  "F722", # Incompatible with jaxtyping
  "E402", # False alarms on autoreload magic in Jupyter Notebooks.
  "ERA001" # Allow commented out code
]
